+++ lib/reader.c
@@ -18,7 +18,6 @@
 
 #include <sys/stat.h>
 #include <stdio.h>
-#include <stdlib.h>
 #include <string.h>
 
 #include "reader.h"
